backend/requirements.txt
django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
whitenoise==6.7.0
